requirements.txt
huggingface_hub 
transformers
langchain==0.0.251
chainlit
python-dotenv
langsmith